Graph visualization interface improved
